+++ src/lib/code-sample-editor.ts
@@ -18,6 +18,7 @@ import {
   customElement,
   css,
   property,
+  internalProperty,
   query,
   PropertyValues,
 } from 'lit-element';
@@ -83,8 +84,27 @@ export class CodeSampleEditor extends LitElement {
   @property({attribute: false})
   files?: SampleFile[];
 
-  // TODO: make a public property/method to select a file
-  @property({attribute: false})
+  /**
+   * The name of the project file that is currently being displayed. Set when
+   * changing tabs. Does not reflect to attribute.
+   */
+  @property()
+  filename?: string;
+
+  /**
+   * If true, don't display the top file-picker. Default: false (visible).
+   */
+  @property({type: Boolean, attribute: 'no-file-picker'})
+  noFilePicker = false;
+
+  /**
+   * If true, display a left-hand-side gutter with line numbers. Default false
+   * (hidden).
+   */
+  @property({type: Boolean, attribute: 'line-numbers'})
+  lineNumbers = false;
+
+  @internalProperty()
   private _currentFileIndex?: number;
 
   private get _currentFile() {
@@ -113,13 +133,16 @@ export class CodeSampleEditor extends LitElement {
     if (changedProperties.has('project')) {
       this._findProjectAndRegister();
     }
-    if (changedProperties.has('files')) {
-      this._currentFileIndex = 0;
+    if (changedProperties.has('files') || changedProperties.has('filename')) {
+      this._currentFileIndex =
+        this.files && this.filename
+          ? this.files.map((f) => f.name).indexOf(this.filename)
+          : 0;
       // TODO(justinfagnani): whyyyy?
       if (this._tabBar) {
         await this._tabBar.updateComplete;
         this._tabBar.activeIndex = -1;
-        this._tabBar.activeIndex = 0;
+        this._tabBar.activeIndex = this._currentFileIndex;
       }
     }
     super.update(changedProperties);
@@ -127,24 +150,27 @@ export class CodeSampleEditor extends LitElement {
 
   render() {
     return html`
-      <mwc-tab-bar
-        .activeIndex=${this._currentFileIndex || 0}
-        @MDCTabBar:activated=${this._tabActivated}
-      >
-        ${this.files?.map((file) => {
-          const label = file.name.substring(file.name.lastIndexOf('/') + 1);
-          return html`<mwc-tab label=${label}></mwc-tab>`;
-        })}
-        ${this.enableAddFile
-          ? html`<mwc-icon-button icon="add"></mwc-icon-button>`
-          : nothing}
-      </mwc-tab-bar>
+      ${this.noFilePicker
+        ? nothing
+        : html` <mwc-tab-bar
+            .activeIndex=${this._currentFileIndex ?? 0}
+            @MDCTabBar:activated=${this._tabActivated}
+          >
+            ${this.files?.map((file) => {
+              const label = file.name.substring(file.name.lastIndexOf('/') + 1);
+              return html`<mwc-tab label=${label}></mwc-tab>`;
+            })}
+            ${this.enableAddFile
+              ? html`<mwc-icon-button icon="add"></mwc-icon-button>`
+              : nothing}
+          </mwc-tab-bar>`}
 
       <codemirror-editor
         .value=${this._currentFile?.content}
         .type=${this._currentFile
           ? mimeTypeToTypeEnum(this._currentFile.contentType)
           : undefined}
+        .lineNumbers=${this.lineNumbers}
         @change=${this._onEdit}
       >
       </codemirror-editor>
@@ -153,6 +179,7 @@ export class CodeSampleEditor extends LitElement {
 
   private _tabActivated(e: CustomEvent<{index: number}>) {
     this._currentFileIndex = e.detail.index;
+    this.filename = this.files?.[this._currentFileIndex].name;
   }
 
   private _findProjectAndRegister() {